Only reference template_fields from unmapped task
